not allowing to fetch, post and delete a post unless logged in. Token authorization

+++ client/src/components/PostDetails.jsx
@@ -4,15 +4,23 @@ import { AiOutlineHeart } from "react-icons/ai"
 
 // Hooks
 import { usePostsContext } from "../hooks/usePostsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 
 
 const PostDetails = ({ post }) => {
     const { dispatch } = usePostsContext()
+    const {user} = useAuthContext()
 
     const handleClick = async () => {
+        if(!user) {
+            return
+        }
         const response = await fetch("http://localhost:5000/api/posts/" + post._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                "Authorization": `Bearer ${user.token}` 
+            }
         })
         const json = await response.json()
